third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageData-expected.txt

Test canvas createImageData()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.createImageData(null) threw exception NotSupportedError: Failed to execute 'createImageData' on 'CanvasRenderingContext2D': The 1st argument provided is either null, or an invalid ImageData object..
PASS imageData.width is 100
PASS imageData.height is 50
PASS imageData.data[32] is 255
PASS sameSizeImageData.width is 100
PASS sameSizeImageData.height is 50
PASS sameSizeImageData.data[32] is 0
PASS imgdata1.data.length is 800
PASS imgdata1.data.length is imgdata2.data.length
PASS imgdata2.data.length is imgdata3.data.length
PASS imgdata3.data.length is imgdata4.data.length
PASS successfullyParsed is true

TEST COMPLETE
